+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/PetstoreImpl.java
@@ -66,16 +66,20 @@ public final class PetstoreImpl extends DomainModel {
 					.column()
 					.caption("Image"),
 				Address.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				Address.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				Address.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				Address.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Address")
@@ -101,16 +105,20 @@ public final class PetstoreImpl extends DomainModel {
 					.caption("Image url")
 					.maximumLength(55),
 				Category.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				Category.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				Category.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				Category.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Category")
@@ -137,16 +145,20 @@ public final class PetstoreImpl extends DomainModel {
 					.nullable(false)
 					.maximumLength(24),
 				ContactInfo.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				ContactInfo.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				ContactInfo.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				ContactInfo.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Contact info")
@@ -176,16 +188,20 @@ public final class PetstoreImpl extends DomainModel {
 					.nullable(false)
 					.maximumLength(30),
 				Tag.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				Tag.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				Tag.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				Tag.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Tag")
@@ -218,16 +234,20 @@ public final class PetstoreImpl extends DomainModel {
 					.caption("Image url")
 					.maximumLength(55),
 				Product.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				Product.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				Product.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				Product.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Product")
@@ -292,16 +312,20 @@ public final class PetstoreImpl extends DomainModel {
 					.nullable(false)
 					.columnHasDefaultValue(true),
 				Item.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				Item.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				Item.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				Item.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Item")
@@ -321,16 +345,20 @@ public final class PetstoreImpl extends DomainModel {
 					.foreignKey()
 					.caption("Item"),
 				TagItem.INSERT_USER.define()
-					.auditInsertUserColumn()
+					.auditColumn()
+					.insertUser()
 					.caption("Insert user"),
 				TagItem.INSERT_TIME.define()
-					.auditInsertTimeColumn()
+					.auditColumn()
+					.insertTime()
 					.caption("Insert time"),
 				TagItem.UPDATE_USER.define()
-					.auditUpdateTimeColumn()
+					.auditColumn()
+					.updateUser()
 					.caption("Update user"),
 				TagItem.UPDATE_TIME.define()
-					.auditUpdateUserColumn()
+					.auditColumn()
+					.updateTime()
 					.caption("Update time"))
 			.caption("Tag item")
 			.build();